requirements.txt
pytest >= 6.2.2
coverage >= 5.4
memory_profiler >= 0.5
flask-restful
flask >=.1
PyPDF2
pymongo
dnspython==2.0.0
google-cloud-language